+++ metrics/interface/src/main/java/org/apache/iotdb/metrics/type/Counter.java
@@ -18,7 +18,7 @@
  */
 package org.apache.iotdb.metrics.type;
 
-public interface Counter extends Metric {
+public interface Counter extends IMetric {
   void inc();
   void inc(long n);
   void dec();
