+++ metrics/interface/src/main/java/org/apache/iotdb/metrics/type/Histogram.java
@@ -18,10 +18,10 @@
  */
 package org.apache.iotdb.metrics.type;
 
-public interface Histogram extends Metric {
+public interface Histogram extends IMetric {
   void update(int value);
   void update(long value);
   long count();
 
-  Snapshot takeSnapshot();
+  HistogramSnapshot takeSnapshot();
 }
