+++ metrics/interface/src/main/java/org/apache/iotdb/metrics/type/Rate.java
@@ -18,5 +18,13 @@
  */
 package org.apache.iotdb.metrics.type;
 
-public interface Metric {
+public interface Rate extends IMetric {
+  long getCount();
+  double getOneMinuteRate();
+  double getMeanRate();
+  double getFiveMinuteRate();
+  double getFifteenMinuteRate();
+
+  void mark();
+  void mark(long n);
 }
